# src/main/resources/api-contracts.yaml
openapi: 3.0.0
info:
  title: Global API Gateway
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development API Gateway
...
paths:
...
  /summoner/{puuid}:
    parameters:
      - in: path
        name: puuid
        schema:
          type: string
        required: true
        description: The puuid of the summoner.
    get:
      summary: Retrieve a summoner by its puuid.
      operationId: getSummonerByPuuid
      tags:
        - Summoner
      responses:
        '200':
          description: Summoner successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummonerDTO'
        '404':
          description: Summoner not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    post:
      summary: Update a summoner.
      operationId: updateSummoner
      tags:
        - Summoner
      responses:
        '200':
          description: Summoner successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummonerDTO'
        '404':
          description: Summoner not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /summoner/{tag}/{name}:
    parameters:
      - in: path
        name: tag
        schema:
          type: string
        required: true
        description: The tag of the summoner.
      - in: path
        name: name
        schema:
          type: string
        required: true
        description: The name of the summoner.
    get:
      summary: Retrieve a summoner by its tag and its name.
      operationId: getSummonerByTagAndName
      tags:
        - Summoner
      responses:
        '200':
          description: Summoner successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummonerDTO'
        '404':
          description: Summoner not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
...
  /match/{puuid}/{pageNb}/{queue}/{role}:
    parameters:
      - in: path
        name: puuid
        schema:
          type: string
        required: true
        description: The puuid of the summoner.
      - in: path
        name: pageNb
        schema:
          type: integer
          format: int32
        required: true
        description: The number of the page.
      - in: path
        name: queue
        schema:
          type: string
        required: true
        description: The requested queue.
      - in: path
        name: role
        schema:
          type: string
        required: true
        description: The requested role.
    get:
      summary: Retrieve all summoner matches of the requested page.
      operationId: getMatches
      tags:
        - Match
      responses:
        '200':
          description: Matches successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /match/count/{puuid}/{queue}/{role}:
    parameters:
      - in: path
        name: puuid
        schema:
          type: string
        required: true
        description: The puuid of the summoner.
      - in: path
        name: queue
        schema:
          type: string
        required: true
        description: The requested queue.
      - in: path
        name: role
        schema:
          type: string
        required: true
        description: The requested role.
    get:
      summary: Retrieve summoner matches count.
      operationId: getMatchesCount
      tags:
        - Match
      responses:
        '200':
          description: Matches count successfully retrieved.
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /match/{puuid}:
    parameters:
      - in: path
        name: puuid
        schema:
          type: string
        required: true
        description: The puuid of the summoner.
    post:
      summary: Update all summoner matches
      operationId: updateMatches
      tags:
        - Match
      responses:
        '200':
          description: Matches successfully updated.
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /board/{boardId}:
    parameters:
      - in: path
        name: boardId
        schema:
          type: string
        required: true
        description: The id of the board.
    get:
      summary: Retrieve a board by its id.
      operationId: getBoard
      tags:
        - Board
      responses:
        '200':
          description: Board successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDTO'
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    post:
      summary: Update a board.
      operationId: updateBoard
      tags:
        - Board
      responses:
        '200':
          description: Board successfully updated.
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    delete:
      summary: Delete a board.
      operationId: deleteBoard
      tags:
        - Board
      responses:
        '200':
          description: Board successfully deleted.
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /board/{boardId}/{name}:
    parameters:
      - in: path
        name: boardId
        schema:
          type: string
        required: true
        description: The id of the board.
      - in: path
        name: name
        schema:
          type: string
        required: true
        description: The new name of the board.
    post:
      summary: Update a board name.
      operationId: updateBoardName
      tags:
        - Board
      responses:
        '200':
          description: Board successfully updated with the new name.
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /board/{boardId}/remove/{puuid}:
    parameters:
      - in: path
        name: boardId
        schema:
          type: string
        required: true
        description: The id of the board.
      - in: path
        name: puuid
        schema:
          type: string
        required: true
        description: The puuid of the summoner.
    post:
      summary: Remove a summoner of a board.
      operationId: removeSummonerOfBoard
      tags:
        - Board
      responses:
        '200':
          description: Summoner successfully removed of the board.
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /board/{boardId}/add/{name}:
    parameters:
      - in: path
        name: boardId
        schema:
          type: string
        required: true
        description: The id of the board.
      - in: path
        name: tag
        schema:
          type: string
        required: true
        description: The tag of the summoner.
      - in: path
        name: name
        schema:
          type: string
        required: true
        description: The name of the summoner.
    post:
      summary: Add a summoner of a board.
      operationId: addSummonerToBoard
      tags:
        - Board
      responses:
        '200':
          description: Summoner successfully added of the board.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummonerDTO'
        '404':
          description: Board not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /board:
    post:
      summary: Create a board.
      operationId: createBoard
      tags:
        - Board
      responses:
        '200':
          description: Board successfully created.
          content:
            application/json:
              schema:
                type: string
                example: "newBoardId"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
components:
  schemas:
    ApiErrorDTO:
      type: object
      description: |
        Represents a standardized error response returned by the API when a request
        cannot be processed successfully. It provides a machine-readable code,
        a human-readable message, and optional detailed information for specific issues.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A unique, application-specific error code.
          example: "INVALID_INPUT"
        message:
          type: string
          description: A human-readable error message that provides a brief explanation of the error.
          example: "The provided 'date' parameter has an invalid format."
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: The name of the field that caused the error.
                example: "date"
              issue:
                type: string
                description: A specific issue with the field.
                example: "must respect the format"
          description: A list of specific validation errors.
...
    SummonerDTO:
      type: object
      properties:
        puuid:
          type: string
          description: Unique identifier for the player.
          example: "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"
        name:
          type: string
          description: The summoner's in-game name.
          example: "SummonerName"
        level:
          type: integer
          format: int32
          description: The summoner's current level.
          example: 123
        icon:
          type: integer
          format: int32
          description: The ID of the summoner icon.
          example: 29
        tag:
          type: string
          description: The summoner's in-game tag.
          example: "#EUW"
        region:
          type: string
          description: The region where the summoner plays.
          example: "euw1"
        updateDate:
          type: string
          format: date-time
          description: The last date and time this summoner's data was updated.
          example: "2025-06-29T22:13:09Z"
        spentTime:
          type: integer
          format: int64
          description: Cumulative time spent by the summoner in game.
          example: 1234567890
        playedSeasonsOrSets:
          type: array
          items:
            type: integer
            format: int32
          description: A set of unique identifiers for seasons or sets the summoner has participated in.
          example: [ 1, 2, 3, 5 ]
      required:
        - puuid
        - name
        - level
        - icon
        - region
        - updateDate
    BoardDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the board.
          example: "board_12345"
        puuids:
          type: array
          items:
            type: string
          description: A list of PUUIDs associated with this board.
          example:
            - "puuid1"
            - "puuid2"
        updateDate:
          type: string
          format: date-time
          description: The last date and time this board's data was updated.
          example: "2025-06-29T23:26:26Z"
        name:
          type: string
          description: The name of the board.
          example: "My Custom Board"
...
    ParticipantDTO:
      type: object
      properties:
        champLevel:
          type: integer
          format: int32
          description: The participant's champion level.
          example: 345
        kills:
          type: integer
          format: int32
          description: Number of kills by the participant.
          example: 12
        assists:
          type: integer
          format: int32
          description: Number of assists by the participant.
          example: 4
        deaths:
          type: integer
          format: int32
          description: Number of deaths by the participant.
          example: 2
        role:
          type: string
          description: The participant's role in the game.
          example: "jungle"
        visionScore:
          type: integer
          format: int32
          description: The participant's vision score.
          example: 40
        minions:
          type: integer
          format: int32
          description: Total minions killed by the participant.
          example: 203
        neutralMinions:
          type: integer
          format: int32
          description: Total neutral minions (jungle monsters) killed by the participant.
          example: 110
        doubleKills:
          type: integer
          format: int32
          description: Number of double kills.
          example: 2
        tripleKills:
          type: integer
          format: int32
          description: Number of triple kills.
          example: 1
        quadraKills:
          type: integer
          format: int32
          description: Number of quadra kills.
          example: 1
        pentaKills:
          type: integer
          format: int32
          description: Number of penta kills.
          example: 1
        physicalDamageDealtToChampions:
          type: integer
          format: int32
          description: Physical damage dealt to champions.
          example: 45020
        magicDamageDealtToChampions:
          type: integer
          format: int32
          description: Magic damage dealt to champions.
          example: 12789
        trueDamageDealtToChampions:
          type: integer
          format: int32
          description: True damage dealt to champions.
          example: 490
        physicalDamageTaken:
          type: integer
          format: int32
          description: Physical damage taken.
          example: 19493
        magicDamageTaken:
          type: integer
          format: int32
          description: Magic damage taken.
          example: 34928
        trueDamageTaken:
          type: integer
          format: int32
          description: True damage taken.
          example: 1342
        golds:
          type: integer
          format: int32
          description: Total gold earned.
          example: 20194
        gameEndedInEarlySurrender:
          type: boolean
          description: Indicates if the game ended in an early surrender for this participant's team.
          example: false
        runes:
          $ref: '#/components/schemas/RunesDTO'
        champion:
          type: string
          description: The ID or name of the champion played.
          example: "122"
        item0:
          type: string
          description: ID or name of item slot 0.
          example: "101"
        item1:
          type: string
          description: ID or name of item slot 1.
          example: "2870"
        item2:
          type: string
          description: ID or name of item slot 2.
          example: "1300"
        item3:
          type: string
          description: ID or name of item slot 3.
          example: "4900"
        item4:
          type: string
          description: ID or name of item slot 4.
          example: "3170"
        item5:
          type: string
          description: ID or name of item slot 5.
          example: "5600"
        ward:
          type: string
          description: ID or name of the ward item.
          example: "120"
        summoner:
          $ref: '#/components/schemas/SummonerDTO'
        summonerSpell1:
          type: string
          description: ID or name of the first summoner spell.
          example: "4"
        summonerSpell2:
          type: string
          description: ID or name of the second summoner spell.
          example: "12"
        placement:
          type: integer
          format: int32
          description: Participant's placement.
          example: 2
        augment1:
          type: string
          description: ID or name of the first augment.
          example: "24"
        augment2:
          type: string
          description: ID or name of the second augment.
          example: "53"
        augment3:
          type: string
          description: ID or name of the third augment.
          example: "28"
        augment4:
          type: string
          description: ID or name of the fourth augment.
          example: "144"
        augment5:
          type: string
          description: ID or name of the fifth augment.
          example: "160"
        augment6:
          type: string
          description: ID or name of the sixth augment.
          example: "120"
    ObjectivesDTO:
      type: object
      properties:
        baronKills:
          type: integer
          format: int32
          description: Number of Baron Nashor kills.
          example: 1
        championKills:
          type: integer
          format: int32
          description: Number of champion kills.
          example: 34
        dragonKills:
          type: integer
          format: int32
          description: Number of Dragon kills.
          example: 4
        hordeKills:
          type: integer
          format: int32
          description: Number of Void Grub kills.
          example: 5
        inhibitorKills:
          type: integer
          format: int32
          description: Number of Inhibitor kills.
          example: 3
        riftHeraldKills:
          type: integer
          format: int32
          description: Number of Rift Herald kills.
          example: 1
        towerKills:
          type: integer
          format: int32
          description: Number of Tower kills.
          example: 7
        atakhanKills:
          type: integer
          format: int32
          description: Number of Atakhan kills.
          example: 1
    TeamDTO:
      type: object
      properties:
        objectives:
          $ref: '#/components/schemas/ObjectivesDTO'
        hasWin:
          type: boolean
          description: Indicates if the team won the game.
          example: true
        id:
          type: string
          description: The ID of the team.
          example: "1"
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantDTO'
          description: List of individual participants in the match.
        bans:
          type: array
          items:
            type: string
          description: List of champion IDs or names banned by this team.
          example:
            - "Yasuo"
            - "Zed"
            - "LeeSin"
            - "Aatrox"
        placement:
          type: integer
          format: int32
          description: The team's final placement in the game.
          example: 1
    MatchDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the match.
          example: "EUW1_694207890"
        version:
          type: string
          description: The game version the match was played on.
          example: "14.12.569.8398"
        datetime:
          type: string
          format: date-time
          description: Date and time when the match started.
          example: "2025-06-30T16:00:00Z"
        duration:
          type: integer
          format: int64
          description: Duration of the game in seconds.
          example: 2450
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamDTO'
          description: List of teams participating in the match.
        remake:
          type: boolean
          description: Indicates if the game was a remake.
          example: false
        queue:
          type: string
          description: The ID or name of the game queue.
          example: "420"
    RunesDTO:
      type: object
      description: Represents all selected runes.
      properties:
        style:
          type: object
          description: Represents the style id of the style category.
          example: "Précision"
          additionalProperties:
            type: integer
        runeCategories:
          type: array
          description: Represents all selected runes categories.
          items:
            $ref: '#/components/schemas/RuneCategoryDTO'
      required:
        - stats
        - runeCategories
    RuneCategoryDTO:
      type: object
      description: Represents a selected rune category.
      properties:
        style:
          type: string
          description: Represents the style id of the style category.
          example: "Précision"
        perks:
          type: array
          description: A list of all selected perks.
          items:
            type: string
      required:
        - style
        - perks